component Pagination is ok
